[N27] Declare uint as uint256 (#77)

+++ contracts/BasePaymaster.sol
@@ -21,9 +21,9 @@ abstract contract BasePaymaster is IPaymaster, Ownable {
         entryPoint = _entryPoint;
     }
 
-    function validatePaymasterUserOp(UserOperation calldata userOp, bytes32 requestId, uint maxCost) external virtual override view returns (bytes memory context);
+    function validatePaymasterUserOp(UserOperation calldata userOp, bytes32 requestId, uint256 maxCost) external virtual override view returns (bytes memory context);
 
-    function postOp(PostOpMode mode, bytes calldata context, uint actualGasCost) external override {
+    function postOp(PostOpMode mode, bytes calldata context, uint256 actualGasCost) external override {
         _requireFromEntrypoint();
         _postOp(mode, context, actualGasCost);
     }
@@ -40,7 +40,7 @@ abstract contract BasePaymaster is IPaymaster, Ownable {
      * @param context - the context value returned by validatePaymasterUserOp
      * @param actualGasCost - actual gas used so far (without this postOp call).
      */
-    function _postOp(PostOpMode mode, bytes calldata context, uint actualGasCost) internal virtual {
+    function _postOp(PostOpMode mode, bytes calldata context, uint256 actualGasCost) internal virtual {
 
         (mode,context,actualGasCost); // unused params
         // subclass must override this method if validatePaymasterUserOp returns a context
@@ -55,7 +55,7 @@ abstract contract BasePaymaster is IPaymaster, Ownable {
         entryPoint.addStake{value:msg.value}(entryPoint.unstakeDelaySec() + extraUnstakeDelaySec);
     }
 
-    function getDeposit() public view returns (uint) {
+    function getDeposit() public view returns (uint256) {
         return entryPoint.balanceOf(address(this));
     }
 
@@ -74,7 +74,7 @@ abstract contract BasePaymaster is IPaymaster, Ownable {
      * @param withdrawAddress the address to send withdrawn value.
      * @param withdrawAmount the amount to withdraw.
      */
-    function withdrawTo(address payable withdrawAddress, uint withdrawAmount) external onlyOwner {
+    function withdrawTo(address payable withdrawAddress, uint256 withdrawAmount) external onlyOwner {
         entryPoint.withdrawTo(withdrawAddress, withdrawAmount);
     }
